ADD using jwt_claims and jwt_optional

+++ src/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import *
 from src.models.item import ItemModel
 from src.models.store import StoreModel
 
@@ -41,11 +41,19 @@ class Item(Resource):
 
         return item.json(), 201  # create success
 
+    @jwt_required
     def delete(self, name):
+        # using JWT claim
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message':'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
 
         if item:
             item.delete_from_db()
+            return {'message':'Item deleted'}
+        return {'message':'Item no found.'}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -62,8 +70,17 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-       return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+
+        if user_id:
+            return {'items':items}, 200
+
+        return {'items': [item['name'] for item in items],
+                'message':'More data available if you log-in'
+                }
 
     def delete(self):
         ItemModel.delete_all_data_from_db()
